+++ src/Misc/layoutbin/en-US/strings.json
@@ -27,10 +27,10 @@
   "AttachFileNotExist": "Can't attach (type:{0} name:{1}) file: {2}. File does not exist.",
   "AttachmentExceededMaximum": "Skipping attachment as it exceeded the maximum allowed size: {0}",
   "AttemptRemoveCredFromConfig": "An unsuccessful attempt was made using git command line to remove \"http.extraheader\" from the git config. Attempting to modify the git config file directly to remove the credential.",
-  "AutoLogonAccountNameDescription":"User account to use for autologon",
-  "AutoLogonAlreadyEnabledWarning" : "WARNING: Autologon is already enabled for the user '{0}'. This setting will be overwritten now.",
-  "AutoLogonConfigurationFailureMessage": "AutoLogon configuration failed on the machine. Refer logs for more details. Invoke run.cmd to run agent without AutoLogon.",
   "AuthenticationType": "authentication type",
+  "AutoLogonAccountNameDescription": "User account to use for autologon",
+  "AutoLogonAlreadyEnabledWarning": "WARNING: Autologon is already enabled for the user '{0}'. This setting will be overwritten now.",
+  "AutoLogonConfigurationFailureMessage": "AutoLogon configuration failed on the machine. Refer logs for more details. Invoke run.cmd to run agent without AutoLogon.",
   "BuildCommandNotFound": "##vso[build.{0}] is not a recognized command for Build command extension. Please reference documentaion (http://go.microsoft.com/fwlink/?LinkId=817296)",
   "BuildDirLastUseTIme": "The last time build directory '{0}' been used is: {1}",
   "BuildNumberRequired": "Build number is required.",
@@ -84,7 +84,7 @@
   "DownloadArtifacts": "Download Artifacts",
   "DownloadArtifactsFailed": "Downloading artifacts failed: {0}",
   "DownloadingTask0": "Downloading task: {0}",
-  "EnableAutoLogon":"enable AutoLogon? (Y/N)",
+  "EnableAutoLogon": "enable AutoLogon? (Y/N)",
   "EnsureJobFinished": "Waiting for current job finish running.",
   "EnsureTasksExist": "Download all required tasks.",
   "EnterValidValueFor0": "Enter a valid value for {0}.",
@@ -323,9 +323,9 @@
   "RegisterAgentSectionHeader": "Register Agent",
   "RenameIndexFileCoberturaFailed": "Renaming '{0}' to '{1}' failed while publishing code coverage files for '{2}'. Inner Exception: '{3}'",
   "Replace": "replace? (Y/N)",
-  "ResultsCommandNotFound": "##vso[results.{0}] is not a recognized command for Task command extension. TODO: DOC aka link",
   "RestartMessage": "Agent configuration for autologon capability is completed. It needs a restart for the changes to take effect.",
   "RestartNow": "Do you want to restart the machine now? (Y/N)",
+  "ResultsCommandNotFound": "##vso[results.{0}] is not a recognized command for Task command extension. TODO: DOC aka link",
   "RetryingRMArtifactCleanUp": "Failed to cleanup Release's artifact directory '{0}' with an exception {1}. Retrying the cleanup of Release artifact directory.",
   "RMArtifactContainerDetailsInvaidError": "The artifact does not have valid container details: {0}",
   "RMArtifactContainerDetailsNotFoundError": "The artifact does not contain container details: {0}",
@@ -435,10 +435,10 @@
   "TotalUploadFiles": "Uploading {0} files",
   "TypeRequiredForTimelineRecord": "Type is required for this new timeline record.",
   "UITestingWarning": "WARNING: Following settings on the machine may affect the UI test runs on the machine.",
-  "UITestingWarning_ScreenSaver": "Screen Saver",
-  "UITestingWarning_ShutdownReason": "Shutdown reason",
   "UITestingWarning_AutoLogonCount": "AutoLogon count",
   "UITestingWarning_LegalNotice": "Legal Notice",
+  "UITestingWarning_ScreenSaver": "Screen Saver",
+  "UITestingWarning_ShutdownReason": "Shutdown reason",
   "UnableResolveArtifactType": "Can't infer artifact type from artifact location: {0}.",
   "UnableToArchiveResults": "Unable to archive the test results: {0}",
   "UnableToParseBuildTrackingConfig0": "Unable to parse the legacy build tracking configuration. A new build directory will be created instead. The previous directory may be left in an unreclaimed state. Legacy configuration contents: {0}",
@@ -462,5 +462,5 @@
   "WindowsLogonAccountNameDescription": "User account to use for the service",
   "WindowsLogonPasswordDescription": "Password for the account {0}",
   "WorkFolderDescription": "work folder",
-  "Y": "Y"  
+  "Y": "Y"
 }
