+++ src/Test/L0/Listener/Configuration/AgentCapabilitiesProviderTestL0.cs
@@ -4,6 +4,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using Xunit;
+using Moq;
+using Microsoft.VisualStudio.Services.Agent.Listener.Configuration;
 
 namespace Microsoft.VisualStudio.Services.Agent.Tests.Listener
 {
@@ -17,6 +19,9 @@ namespace Microsoft.VisualStudio.Services.Agent.Tests.Listener
             using (var hc = new TestHostContext(this))
             using (var tokenSource = new CancellationTokenSource())
             {
+                Mock<IConfigurationManager> configurationManager = new Mock<IConfigurationManager>();
+                hc.SetSingleton<IConfigurationManager>(configurationManager.Object);
+                
                 // Arrange
                 var provider = new AgentCapabilitiesProvider();
                 provider.Initialize(hc);
@@ -32,5 +37,35 @@ namespace Microsoft.VisualStudio.Services.Agent.Tests.Listener
                 Assert.Equal("IAmAgent007", agentNameCapability.Value);
             }
         }
+
+        [Fact]
+        [Trait("Level", "L0")]
+        [Trait("Category", "Agent")]
+        public async void TestInteractiveSessionCapability()
+        {
+            using (var hc = new TestHostContext(this))
+            using (var tokenSource = new CancellationTokenSource())
+            {
+                Mock<IConfigurationManager> configurationManager = new Mock<IConfigurationManager>();
+                hc.SetSingleton<IConfigurationManager>(configurationManager.Object);
+
+                configurationManager.Setup(x => x.IsServiceConfigured()).Returns(false);
+
+                // Arrange
+                var provider = new AgentCapabilitiesProvider();
+                provider.Initialize(hc);
+                var settings = new AgentSettings() { AgentName = "IAmAgent007" };
+
+                // Act
+                List<Capability> capabilities = await provider.GetCapabilitiesAsync(settings, tokenSource.Token);
+
+                // Assert
+                Assert.NotNull(capabilities);
+                Capability iSessionCapability = capabilities.SingleOrDefault(x => string.Equals(x.Name, "InteractiveSession", StringComparison.Ordinal));
+                Assert.NotNull(iSessionCapability);
+                
+                Assert.True(iSessionCapability.Value.Equals("true", StringComparison.OrdinalIgnoreCase));
+            }
+        }        
     }
 }
